requirements.txt
instagrapi==1.16.31
Pillow==9.4.0
praw==7.6.1
requests==2.28.2

RedDownloader~=4.2.0